# POC for a way to speed up search tests.