@reupen 2019-10-11
Could we use the `class` fixture scope to do this?

https://docs.pytest.org/en/latest/fixture.html#scope-sharing-a-fixture-instance-across-tests-in-a-class-module-or-session