@reupen 2019-10-14
It's a good observation about that test.

Just a bit apprehensive about using internals for this.

I had a look locally and it seems like a lot of time is wasted syncing objects to ES individually.

For example, this yields a similar improvement (and can still be parallelised):

```python
        with disable_search_signal_receivers(Order), \
                disable_search_signal_receivers(Company), \
                disable_search_signal_receivers(Contact):
            order_with_multiple_refunds = OrderPaidFactory()
            ApprovedRefundFactory(
                order=order_with_multiple_refunds,
                requested_amount=order_with_multiple_refunds.total_cost / 5,
            )
            ApprovedRefundFactory(
                order=order_with_multiple_refunds,
                requested_amount=order_with_multiple_refunds.total_cost / 4,
            )

            for factory_ in factories:
                factory_.create_batch(2)

        sync_all_models()

        setup_es.indices.refresh()
```

I think we could come up with a smarter fixture for tests like these that syncs in bulk (and hence works more generically).