@benpbenp 2019-10-14
@reupen What do you mean internals? We could of course combine your optimisation with mine for even faster tests.