@reupen 2019-10-14
> What do you mean internals? 

Anything prefixed with `_` is usually considered internal. For example, the entire `_pytest` package is internal.