@benpbenp 2019-10-14
Oh I see, the monkeypatch one is. Perhaps that part of it could be done another way, but the main point is the class-scoped data fixture.