@benpbenp 2019-10-15
@reupen see the alternative way of implementing synchronous_on_commit_for_class that I just pushed, I believe that should work the same as using the monkeypatch fixture.