@currycoder 2019-10-23
Some of the ideas here have made it in to other PRs which improve speed of the test suite.  Closing this for now.